+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add some extra global accuracies (precision, recall, f1) to report (#119)
 - Add option `calc_marker_params.on_missing_images` to be able to ignore errors in the
   calculation of images if needed (#125, #126)
+- Filter away rasterio logging for extrasamples (#127)
 
 ## 0.2 (2024-06-17)